requirements.txt
asgiref==3.5.0
boto3==1.21.24
botocore==1.24.24
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
django-allauth==0.49.0
django-cors-headers==3.11.0
django-mysql==4.5.0
django-ninja==0.17.0
django-storages==1.12.3
filelock==3.6.0
idna==3.3
jmespath==1.0.0
joblib==1.1.0
oauthlib==3.2.0
platformdirs==2.5.1
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.4
